Move new ESky2 Protocol to replace old ESky150 protocol.
