+++ src/protocol/protocol.h
@@ -37,7 +37,6 @@ PROTODEF(PROTOCOL_HM830, NRF24L01, TAERG, HM830_Cmds, "HM830")
 PROTODEF(PROTOCOL_KN, NRF24L01, TAERG, KN_Cmds, "KN")
 PROTODEF(PROTOCOL_ESKY150, NRF24L01, TAERG, ESKY150_Cmds, "ESky150")
 PROTODEF(PROTOCOL_ESKY, NRF24L01, AETRG, ESKY_Cmds, "ESky")
-PROTODEF(PROTOCOL_ESKY2, NRF24L01, TAERG, ESky2_Cmds, "ESky2")  
 PROTODEF(PROTOCOL_BLUEFLY, NRF24L01, AETRG, BlueFly_Cmds, "BlueFly")
 //PROTODEF(PROTOCOL_NE260, NRF24L01, AETRG, NE260_Cmds, "NE260")
 PROTODEF(PROTOCOL_CX10, NRF24L01, AETRG, CX10_Cmds, "CX10")
